+++ src/Snapshot.php
@@ -67,13 +67,13 @@ final class Snapshot implements JsonSerializable
      */
     public function jsonSerialize(): array
     {
-        return $this->stats();
+        return $this->toArray();
     }
 
     /**
      * @return SnapshotStat
      */
-    public function stats(): array
+    public function toArray(): array
     {
         return [
             'timestamp' => $this->timestamp->format("Y-m-d\TH:i:s.uP"),